fix(dataproject): remove PrivateAssets attribute from project reference
The PrivateAssets="All" attribute was removed from the ProjectReference in RA.Utilities.Data.EntityFramework.csproj. This change allows the referenced RA.Utilities.Data.Abstractions project to be exposed transitively to projects that depend on RA.Utilities.Data.EntityFramework. This may be necessary for scenarios where the abstractions are needed by downstream projects or when the encapsulation provided by PrivateAssets is not required.