chore(docs): update badges and project version
Added logos to NuGet and license badges in README files for better visual identification. Updated the project version in RA.Utilities.Core.Exceptions.csproj from a release candidate to the final version (10.0.0). Expanded documentation in the RA.Utilities.Core.Exceptions README with sections on best practices, how it works, and additional examples to improve user guidance and clarity on package usage. These enhancements aim to improve readability and usability of the documentation for users and developers.