docs(api): update badges and links in middleware documentation
Revised the badge section in the RA.Utilities.Api.Middlewares documentation to correct and enhance badge display. Added label to the NuGet version badge for clarity, updated the GitHub license badge label, and adjusted the link for the Publish NuGet badge to reflect the latest action run. Removed the redundant Documentation badge as it was potentially outdated or unnecessary. These changes improve the readability and accuracy of the badge information presented to users.